refactoring the directory structure
