+++ lib/components/common/message_area.dart
