+++ lib/components/desktop/main_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:morelia_client_flutter/components/communication_page.dart';
-import 'package:morelia_client_flutter/desktop/vertical_menu_bar.dart';
+import 'package:morelia_client_flutter/components/common/communication_page.dart';
+import '../desktop/vertical_menu_bar.dart';
 
-import '../components/chat_list.dart';
+import '../common/chat_list.dart';
 
 final currentChatPage =
     StateProvider<Widget>((ref) => const Center(child: Text("Select chat")));
@@ -21,7 +21,7 @@ class DesktopMainPage extends ConsumerWidget {
                 CommunicationPage(uuid: uuid));
     return Scaffold(
       body: (Row(children: [
-        VerticalMenuBar(),
+        const VerticalMenuBar(),
         const Expanded(flex: 3, child: ChatList()),
         Expanded(flex: 6, child: ref.watch(currentChatPage))
       ])),
