+++ lib/components/desktop/vertical_menu_bar.dart
