+++ lib/components/mobile/chats_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:morelia_client_flutter/mobile/mobile_nav_bar.dart';
+import 'package:morelia_client_flutter/components/mobile/mobile_nav_bar.dart';
 
-import '../components/chat_list.dart';
+import '../common/chat_list.dart';
 
 class MobileChatsPage extends ConsumerWidget {
   const MobileChatsPage({Key? key}) : super(key: key);
