+++ lib/main.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:morelia_client_flutter/desktop/main_page.dart';
-import 'package:morelia_client_flutter/mobile/chats_page.dart';
+import 'package:morelia_client_flutter/components/desktop/main_page.dart';
+import 'package:morelia_client_flutter/components/mobile/chats_page.dart';
 import 'package:morelia_client_flutter/modules/platform_const.dart';
 import 'package:morelia_client_flutter/modules/theme_manager.dart'
     show currentTheme;
 
-import 'components/communication_page.dart';
+import 'components/common/communication_page.dart';
 
 void main() {
   runApp(const ProviderScope(child: MoreliaApp()));
