+++ test/components/communication_page_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:morelia_client_flutter/components/communication_page.dart';
-import 'package:morelia_client_flutter/components/message_area.dart';
+import 'package:morelia_client_flutter/components/common/communication_page.dart';
+import 'package:morelia_client_flutter/components/common/message_area.dart';
 
 void main() {
   testWidgets('Test widget "CommunicationPage" create',
